+++ src/__flowtests__/useForm-flowtest.js
@@ -2,9 +2,8 @@
 /* eslint-disable no-console */
 
 import * as React from 'react';
-import { Button, View } from 'react-native';
 import { useForm } from '..';
-import { FieldSwitch, FieldTextInput } from '../bridge/components/react-native';
+import { FieldCheckbox, FieldText } from '../__tests__/samples';
 
 import type { $Optional } from '..';
 
@@ -26,13 +25,13 @@ function Screen() {
     },
   });
   return (
-    <View>
-      <FieldTextInput field={getField('username')} />
-      <FieldTextInput field={getField('password')} secureTextEntry />
+    <form>
+      <FieldText field={getField('username')} />
+      <FieldText field={getField('password')} secureTextEntry />
       {/* $FlowExpectedError */}
-      <FieldTextInput field={getField('noexist')} />
-      <Button title="Submit" onPress={handleSubmit} />
-    </View>
+      <FieldText field={getField('noexist')} />
+      <input type="submit" onPress={handleSubmit} />
+    </form>
   );
 }
 <Screen />;
@@ -52,11 +51,11 @@ function DefaultsScreen() {
     },
   });
   return (
-    <View>
-      <FieldTextInput field={getField('username')} />
-      <FieldTextInput field={getField('password')} secureTextEntry />
-      <Button title="Submit" onPress={handleSubmit} />
-    </View>
+    <form>
+      <FieldText field={getField('username')} />
+      <FieldText field={getField('password')} secureTextEntry />
+      <input type="submit" onPress={handleSubmit} />
+    </form>
   );
 }
 <DefaultsScreen />;
@@ -67,7 +66,7 @@ function DefaultsBrokenUndefScreen() {
     defaultValues: { username: 'anon', __password: 'foo' },
     onSubmit: values => console.log(values),
   });
-  return <View />;
+  return <form />;
 }
 <DefaultsBrokenUndefScreen />;
 
@@ -77,7 +76,7 @@ function DefaultsBrokenMissingScreen() {
     defaultValues: { password: 'foo' },
     onSubmit: values => console.log(values),
   });
-  return <View />;
+  return <form />;
 }
 <DefaultsBrokenMissingScreen />;
 
@@ -91,7 +90,7 @@ function PristineScreen() {
     pristineValues: { username: 'anon' },
     onSubmit: values => console.log(values),
   });
-  return <View />;
+  return <form />;
 }
 <PristineScreen />;
 
@@ -108,11 +107,11 @@ function PristineModelScreen({ user }: { user: User }) {
     onSubmit: values => console.log(values),
   });
   return (
-    <View>
-      <FieldTextInput field={getField('username')} />
-      <FieldTextInput field={getField('password')} secureTextEntry />
-      <FieldSwitch field={getField('enabled')} />
-    </View>
+    <form>
+      <FieldText field={getField('username')} />
+      <FieldText field={getField('password')} secureTextEntry />
+      <FieldCheckbox field={getField('enabled')} />
+    </form>
   );
 }
 <PristineModelScreen
@@ -125,6 +124,6 @@ function PristineBrokenScreen() {
     pristineValues: { username: 2, other: 'yes' },
     onSubmit: values => console.log(values),
   });
-  return <View />;
+  return <form />;
 }
 <PristineBrokenScreen />;