+++ src/__tests__/samples/index.js
@@ -1,0 +1,6 @@
+// @flow
+
+export { default as FieldCheckbox } from './FieldCheckbox';
+export { default as FieldText } from './FieldText';
+export { default as FieldTextNullable } from './FieldTextNullable';
+export { default as createValidator } from './create-validator';