+++ src/bridge/components/react-native/index.js
@@ -1,5 +1,0 @@
-// @flow
-
-export { default as FieldSwitch } from './FieldSwitch';
-export { default as FieldTextInput } from './FieldTextInput';
-export { default as FieldTextInputNullable } from './FieldTextInputNullable';