+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- **BC Break**: Remove bridges for React Native and validate.js.
+
 ## 0.2.3
 
 - Fix usage of `void` in TS types (replace with `undefined`).